+++ src/layout.rs
@@ -1,6 +1,6 @@
 //! Layout management.
 
-use crate::action::{Action, SequenceEvent};
+use crate::action::{Action, HoldTapConfig, SequenceEvent};
 use crate::key_code::KeyCode;
 use arraydeque::ArrayDeque;
 use heapless::consts::U64;
@@ -13,16 +13,21 @@ use State::*;
 /// The first level correspond to the layer, the two others to the
 /// switch matrix.  For example, `layers[1][2][3]` correspond to the
 /// key i=2, j=3 on the layer 1.
-pub type Layers = &'static [&'static [&'static [Action]]];
+pub type Layers<T = core::convert::Infallible> = &'static [&'static [&'static [Action<T>]]];
+
+type Stack = ArrayDeque<[Stacked; 16], arraydeque::behavior::Wrapping>;
 
 /// The layout manager. It takes `Event`s and `tick`s as input, and
 /// generate keyboard reports.
-pub struct Layout {
-    layers: Layers,
+pub struct Layout<T = core::convert::Infallible>
+where
+    T: 'static,
+{
+    layers: Layers<T>,
     default_layer: usize,
-    states: Vec<State, U64>,
-    waiting: Option<WaitingState>,
-    stacked: ArrayDeque<[Stacked; 16], arraydeque::behavior::Wrapping>,
+    states: Vec<State<T>, U64>,
+    waiting: Option<WaitingState<T>>,
+    stacked: Stack,
     sequenced: ArrayDeque<[SequenceEvent; 16], arraydeque::behavior::Wrapping>,
 }
 
@@ -66,32 +71,68 @@ impl Event {
             }
         }
     }
+
+    /// Returns `true` if the event is a key press.
+    pub fn is_press(self) -> bool {
+        match self {
+            Event::Press(..) => true,
+            Event::Release(..) => false,
+        }
+    }
+
+    /// Returns `true` if the event is a key release.
+    pub fn is_release(self) -> bool {
+        match self {
+            Event::Release(..) => true,
+            Event::Press(..) => false,
+        }
+    }
 }
 
-/// The various states used internally by Keyberon
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-enum State {
-    /// Normal keys
-    NormalKey {
-        /// Keycode
-        keycode: KeyCode,
-        /// Coordinates in the matrix
-        coord: (u8, u8),
-    },
-    /// Layer modifier keys
-    LayerModifier {
-        /// Value
-        value: usize,
-        /// Coordinates of this layer modifier key in the matrix
-        coord: (u8, u8),
-    },
-    /// Fake key event for sequences
-    FakeKey {
-        /// The key to everything!
-        keycode: KeyCode,
-    },
+/// Event from custom action.
+#[derive(Debug, PartialEq, Eq)]
+pub enum CustomEvent<T: 'static> {
+    /// No custom action.
+    NoEvent,
+    /// The given custom action key is pressed.
+    Press(&'static T),
+    /// The given custom action key is released.
+    Release(&'static T),
 }
-impl State {
+impl<T> CustomEvent<T> {
+    /// Update an event according to a new event.
+    ///
+    ///The event can only be modified in the order `NoEvent < Press <
+    /// Release`
+    fn update(&mut self, e: Self) {
+        use CustomEvent::*;
+        match (&e, &self) {
+            (Release(_), NoEvent) | (Release(_), Press(_)) => *self = e,
+            (Press(_), NoEvent) => *self = e,
+            _ => (),
+        }
+    }
+}
+impl<T> Default for CustomEvent<T> {
+    fn default() -> Self {
+        CustomEvent::NoEvent
+    }
+}
+
+#[derive(Debug, Eq, PartialEq)]
+enum State<T: 'static> {
+    NormalKey { keycode: KeyCode, coord: (u8, u8) },
+    LayerModifier { value: usize, coord: (u8, u8) },
+    Custom { value: &'static T, coord: (u8, u8) },
+    FakeKey { keycode: KeyCode, }, // Fake key event for sequences
+}
+impl<T> Copy for State<T> {}
+impl<T> Clone for State<T> {
+    fn clone(&self) -> Self {
+        *self
+    }
+}
+impl<T: 'static> State<T> {
     fn keycode(&self) -> Option<KeyCode> {
         match self {
             NormalKey { keycode, .. } => Some(*keycode),
@@ -108,8 +149,13 @@ impl State {
             _ => Some(*self),
         }
     }
-    fn seq_release(&self, kc: KeyCode) -> Option<Self> {
+    fn release(&self, c: (u8, u8), custom: &mut CustomEvent<T>) -> Option<Self> {
         match *self {
+            NormalKey { coord, .. } | LayerModifier { coord, .. } if coord == c => None,
+            Custom { value, coord } if coord == c => {
+                custom.update(CustomEvent::Release(value));
+                None
+            },
             FakeKey { keycode, .. } if keycode == kc => None,
             _ => Some(*self),
         }
@@ -123,16 +169,55 @@ impl State {
 }
 
 #[derive(Debug, Copy, Clone)]
-struct WaitingState {
+struct WaitingState<T: 'static> {
     coord: (u8, u8),
     timeout: u16,
-    hold: &'static Action,
-    tap: &'static Action,
+    delay: u16,
+    hold: &'static Action<T>,
+    tap: &'static Action<T>,
+    config: HoldTapConfig,
 }
-impl WaitingState {
-    fn tick(&mut self) -> bool {
+enum WaitingAction {
+    Hold,
+    Tap,
+    NoOp,
+}
+impl<T> WaitingState<T> {
+    fn tick(&mut self, stacked: &Stack) -> WaitingAction {
         self.timeout = self.timeout.saturating_sub(1);
-        self.timeout == 0
+        match self.config {
+            HoldTapConfig::Default => (),
+            HoldTapConfig::HoldOnOtherKeyPress => {
+                if stacked.iter().any(|s| s.event.is_press()) {
+                    return WaitingAction::Hold;
+                }
+            }
+            HoldTapConfig::PermissiveHold => {
+                for (x, s) in stacked.iter().enumerate() {
+                    if s.event.is_press() {
+                        let (i, j) = s.event.coord();
+                        let target = Event::Release(i, j);
+                        if stacked.iter().skip(x + 1).any(|s| s.event == target) {
+                            return WaitingAction::Hold;
+                        }
+                    }
+                }
+            }
+        }
+        if let Some(&Stacked { since, .. }) = stacked
+            .iter()
+            .find(|s| self.is_corresponding_release(&s.event))
+        {
+            if self.timeout >= self.delay - since {
+                WaitingAction::Tap
+            } else {
+                WaitingAction::Hold
+            }
+        } else if self.timeout == 0 {
+            WaitingAction::Hold
+        } else {
+            WaitingAction::NoOp
+        }
     }
     fn is_corresponding_release(&self, event: &Event) -> bool {
         matches!(event, Event::Release(i, j) if (*i, *j) == self.coord)
@@ -155,9 +240,9 @@ impl Stacked {
     }
 }
 
-impl Layout {
+impl<T: 'static> Layout<T> {
     /// Creates a new `Layout` object.
-    pub fn new(layers: Layers) -> Self {
+    pub fn new(layers: Layers<T>) -> Self {
         Self {
             layers,
             default_layer: 0,
@@ -171,41 +256,45 @@ impl Layout {
     pub fn keycodes<'a>(&'a self) -> impl Iterator<Item = KeyCode> + 'a {
         self.states.iter().filter_map(State::keycode)
     }
-    fn waiting_into_hold(&mut self) {
+    fn waiting_into_hold(&mut self) -> CustomEvent<T> {
         if let Some(w) = &self.waiting {
             let hold = w.hold;
             let coord = w.coord;
             self.waiting = None;
-            self.do_action(hold, coord, 0);
+            self.do_action(hold, coord, 0)
+        } else {
+            CustomEvent::NoEvent
         }
     }
-    fn waiting_into_tap(&mut self) {
+    fn waiting_into_tap(&mut self) -> CustomEvent<T> {
         if let Some(w) = &self.waiting {
             let tap = w.tap;
             let coord = w.coord;
             self.waiting = None;
-            self.do_action(tap, coord, 0);
+            self.do_action(tap, coord, 0)
+        } else {
+            CustomEvent::NoEvent
         }
     }
     /// A time event.
     ///
     /// This method must be called regularly, typically every millisecond.
     ///
-    /// Returns an iterator on the current key code state.
-    pub fn tick<'a>(&'a mut self) -> impl Iterator<Item = KeyCode> + 'a {
+    /// Returns the corresponding `CustomEvent`, allowing to manage
+    /// custom actions thanks to the `Action::Custom` variant.
+    pub fn tick<'a>(&'a mut self) -> CustomEvent<T> {
         self.states = self.states.iter().filter_map(State::tick).collect();
         self.stacked.iter_mut().for_each(Stacked::tick);
         match &mut self.waiting {
-            Some(w) => {
-                if w.tick() {
-                    self.waiting_into_hold();
-                }
-            }
-            None => {
-                if let Some(s) = self.stacked.pop_front() {
-                    self.unstack(s);
-                }
-            }
+            Some(w) => match w.tick(&self.stacked) {
+                WaitingAction::Hold => self.waiting_into_hold(),
+                WaitingAction::Tap => self.waiting_into_tap(),
+                WaitingAction::NoOp => CustomEvent::NoEvent,
+            },
+            None => match self.stacked.pop_front() {
+                Some(s) => self.unstack(s),
+                None => CustomEvent::NoEvent,
+            },
         }
         self.process_sequences();
         self.keycodes()
@@ -301,41 +390,32 @@ impl Layout {
             }
         }
     }
-    fn unstack(&mut self, stacked: Stacked) {
+    fn unstack(&mut self, stacked: Stacked) -> CustomEvent<T> {
         use Event::*;
         match stacked.event {
             Release(i, j) => {
+                let mut custom = CustomEvent::NoEvent;
                 self.states = self
                     .states
                     .iter()
-                    .filter_map(|s| s.release((i, j)))
-                    .collect()
+                    .filter_map(|s| s.release((i, j), &mut custom))
+                    .collect();
+                custom
             }
             Press(i, j) => {
                 let action = self.press_as_action((i, j), self.current_layer());
-                self.do_action(action, (i, j), stacked.since);
+                self.do_action(action, (i, j), stacked.since)
             }
         }
     }
-    /// A key event.
-    ///
-    /// Returns an iterator on the current key code state.
-    pub fn event<'a>(&'a mut self, event: Event) -> impl Iterator<Item = KeyCode> + 'a {
+    /// Register a key event.
+    pub fn event<'a>(&'a mut self, event: Event) {
         if let Some(stacked) = self.stacked.push_back(event.into()) {
             self.waiting_into_hold();
             self.unstack(stacked);
         }
-        if self
-            .waiting
-            .as_ref()
-            .map(|w| w.is_corresponding_release(&event))
-            .unwrap_or(false)
-        {
-            self.waiting_into_tap();
-        }
-        self.keycodes()
     }
-    fn press_as_action(&self, coord: (u8, u8), layer: usize) -> &'static Action {
+    fn press_as_action(&self, coord: (u8, u8), layer: usize) -> &'static Action<T> {
         use crate::action::Action::*;
         let action = self
             .layers
@@ -354,43 +434,47 @@ impl Layout {
             Some(action) => action,
         }
     }
-    fn do_action(&mut self, action: &Action, coord: (u8, u8), delay: u16) {
+    fn do_action(
+        &mut self,
+        action: &'static Action<T>,
+        coord: (u8, u8),
+        delay: u16,
+    ) -> CustomEvent<T> {
         assert!(self.waiting.is_none());
         use Action::*;
-        match *action {
+        match action {
             NoOp | Trans => (),
-            HoldTap { timeout, hold, tap } => {
-                let waiting = WaitingState {
+            &HoldTap {
+                timeout,
+                hold,
+                tap,
+                config,
+                ..
+            } => {
+                let waiting: WaitingState<T> = WaitingState {
                     coord,
-                    timeout: timeout.saturating_sub(delay),
+                    timeout,
+                    delay,
                     hold,
                     tap,
+                    config,
                 };
                 self.waiting = Some(waiting);
-                if let Some(Stacked { since, .. }) = self
-                    .stacked
-                    .iter()
-                    .find(|s| waiting.is_corresponding_release(&s.event))
-                {
-                    if timeout >= delay - since {
-                        self.waiting_into_tap();
-                    } else {
-                        self.waiting_into_hold();
-                    }
-                }
             }
-            KeyCode(keycode) => {
+            &KeyCode(keycode) => {
                 let _ = self.states.push(NormalKey { coord, keycode });
             }
-            MultipleKeyCodes(v) => {
+            &MultipleKeyCodes(v) => {
                 for &keycode in v {
                     let _ = self.states.push(NormalKey { coord, keycode });
                 }
             }
-            MultipleActions(v) => {
+            &MultipleActions(v) => {
+                let mut custom = CustomEvent::NoEvent;
                 for action in v {
-                    self.do_action(action, coord, delay);
+                    custom.update(self.do_action(action, coord, delay));
                 }
+                return custom;
             }
             Sequence { events } => {
                 // Copy the first chunk of `SequenceEvent`s into the `sequenced` ArrayDeque
@@ -445,15 +529,21 @@ impl Layout {
                     }
                 }
             }
-            Layer(value) => {
+            &Layer(value) => {
                 let _ = self.states.push(LayerModifier { value, coord });
             }
             DefaultLayer(value) => {
-                if value < self.layers.len() {
-                    self.default_layer = value
+                if *value < self.layers.len() {
+                    self.default_layer = *value
+                }
+            }
+            Custom(value) => {
+                if self.states.push(State::Custom { value, coord }).is_ok() {
+                    return CustomEvent::Press(value);
                 }
             }
         }
+        CustomEvent::NoEvent
     }
     fn current_layer(&self) -> usize {
         let mut iter = self.states.iter().filter_map(State::get_layer);
@@ -471,15 +561,16 @@ impl Layout {
 #[cfg(test)]
 mod test {
     extern crate std;
-    use super::{Event::*, Layers, Layout};
+    use super::{Event::*, Layers, Layout, *};
     use crate::action::Action::*;
     use crate::action::SequenceEvent;
+    use crate::action::HoldTapConfig;
     use crate::action::{k, l, m};
     use crate::key_code::KeyCode;
     use crate::key_code::KeyCode::*;
     use std::collections::BTreeSet;
 
-    //#[track_caller]
+    #[track_caller]
     fn assert_keys(expected: &[KeyCode], iter: impl Iterator<Item = KeyCode>) {
         let expected: BTreeSet<_> = expected.iter().copied().collect();
         let tested = iter.collect();
@@ -487,37 +578,196 @@ mod test {
     }
 
     #[test]
-    fn test() {
+    fn basic_hold_tap() {
         static LAYERS: Layers = &[
             &[&[
                 HoldTap {
                     timeout: 200,
                     hold: &l(1),
                     tap: &k(Space),
+                    config: HoldTapConfig::Default,
+                    tap_hold_interval: 0,
                 },
                 HoldTap {
                     timeout: 200,
                     hold: &k(LCtrl),
                     tap: &k(Enter),
+                    config: HoldTapConfig::Default,
+                    tap_hold_interval: 0,
                 },
             ]],
             &[&[Trans, m(&[LCtrl, Enter])]],
         ];
         let mut layout = Layout::new(LAYERS);
-        assert_keys(&[], layout.tick());
-        assert_keys(&[], layout.event(Press(0, 1)));
-        assert_keys(&[], layout.tick());
-        assert_keys(&[], layout.event(Press(0, 0)));
-        assert_keys(&[], layout.tick());
-        assert_keys(&[], layout.event(Release(0, 0)));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Release(0, 0));
         for _ in 0..197 {
-            assert_keys(&[], layout.tick());
+            assert_eq!(CustomEvent::NoEvent, layout.tick());
+            assert_keys(&[], layout.keycodes());
         }
-        assert_keys(&[LCtrl], layout.tick());
-        assert_keys(&[LCtrl, Space], layout.tick());
-        assert_keys(&[LCtrl], layout.tick());
-        assert_keys(&[LCtrl], layout.event(Release(0, 1)));
-        assert_keys(&[], layout.tick());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LCtrl], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LCtrl], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LCtrl, Space], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LCtrl], layout.keycodes());
+        layout.event(Release(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+    }
+
+    #[test]
+    fn hold_tap_interleaved_timeout() {
+        static LAYERS: Layers = &[&[&[
+            HoldTap {
+                timeout: 200,
+                hold: &k(LAlt),
+                tap: &k(Space),
+                config: HoldTapConfig::Default,
+                tap_hold_interval: 0,
+            },
+            HoldTap {
+                timeout: 20,
+                hold: &k(LCtrl),
+                tap: &k(Enter),
+                config: HoldTapConfig::Default,
+                tap_hold_interval: 0,
+            },
+        ]]];
+        let mut layout = Layout::new(LAYERS);
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 1));
+        for _ in 0..15 {
+            assert_eq!(CustomEvent::NoEvent, layout.tick());
+            assert_keys(&[], layout.keycodes());
+        }
+        layout.event(Release(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[Space], layout.keycodes());
+        for _ in 0..10 {
+            assert_eq!(CustomEvent::NoEvent, layout.tick());
+            assert_keys(&[Space], layout.keycodes());
+        }
+        layout.event(Release(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[Space, LCtrl], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LCtrl], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+    }
+
+    #[test]
+    fn hold_on_press() {
+        static LAYERS: Layers = &[&[&[
+            HoldTap {
+                timeout: 200,
+                hold: &k(LAlt),
+                tap: &k(Space),
+                config: HoldTapConfig::HoldOnOtherKeyPress,
+                tap_hold_interval: 0,
+            },
+            k(Enter),
+        ]]];
+        let mut layout = Layout::new(LAYERS);
+
+        // Press another key before timeout
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt, Enter], layout.keycodes());
+        layout.event(Release(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[Enter], layout.keycodes());
+        layout.event(Release(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+
+        // Press another key after timeout
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 0));
+        for _ in 0..200 {
+            assert_eq!(CustomEvent::NoEvent, layout.tick());
+            assert_keys(&[], layout.keycodes());
+        }
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt], layout.keycodes());
+        layout.event(Press(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt, Enter], layout.keycodes());
+        layout.event(Release(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[Enter], layout.keycodes());
+        layout.event(Release(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+    }
+
+    #[test]
+    fn permissive_hold() {
+        static LAYERS: Layers = &[&[&[
+            HoldTap {
+                timeout: 200,
+                hold: &k(LAlt),
+                tap: &k(Space),
+                config: HoldTapConfig::PermissiveHold,
+                tap_hold_interval: 0,
+            },
+            k(Enter),
+        ]]];
+        let mut layout = Layout::new(LAYERS);
+
+        // Press and release another key before timeout
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Release(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt, Enter], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LAlt], layout.keycodes());
+        layout.event(Release(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
     }
 
     #[test]
@@ -527,15 +777,42 @@ mod test {
             &[&[Trans, k(E)]],
         ];
         let mut layout = Layout::new(LAYERS);
-        assert_keys(&[], layout.tick());
-        assert_keys(&[], layout.event(Press(0, 0)));
-        assert_keys(&[LShift], layout.tick());
-        assert_keys(&[LShift], layout.event(Press(0, 1)));
-        assert_keys(&[LShift, E], layout.tick());
-        assert_keys(&[LShift, E], layout.event(Release(0, 1)));
-        assert_keys(&[LShift, E], layout.event(Release(0, 0)));
-        assert_keys(&[LShift], layout.tick());
-        assert_keys(&[], layout.tick());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+        layout.event(Press(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LShift], layout.keycodes());
+        layout.event(Press(0, 1));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LShift, E], layout.keycodes());
+        layout.event(Release(0, 1));
+        layout.event(Release(0, 0));
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[LShift], layout.keycodes());
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+    }
+
+    #[test]
+    fn custom() {
+        static LAYERS: Layers<u8> = &[&[&[Action::Custom(42)]]];
+        let mut layout = Layout::new(LAYERS);
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+
+        // Custom event
+        layout.event(Press(0, 0));
+        assert_eq!(CustomEvent::Press(&42), layout.tick());
+        assert_keys(&[], layout.keycodes());
+
+        // nothing more
+        assert_eq!(CustomEvent::NoEvent, layout.tick());
+        assert_keys(&[], layout.keycodes());
+
+        // release custom
+        layout.event(Release(0, 0));
+        assert_eq!(CustomEvent::Release(&42), layout.tick());
+        assert_keys(&[], layout.keycodes());
     }
 
     #[test]